Add anchor retry logic
The version of Flatdoc currently in use doesn't move to the anchor tag after
generating the anchor tag. There was also no obvious Runner on complete
handler to hook into. Thus sometimes loading the documentation with an
anchor tag ready to scroll to does not result in scrolling.

This fixes the issue by adding retry logic with exponential backoff. The
default settings will retry jumping to the anchor if the user has not
yet scrolled for about 1.7 minutes and 100ms after page load at the
least.